refactored reducers to separate files
